refactor: change package struct
