+++ kopring/src/main/resources/mapper/member/member.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.elixter.kopring.mapper.MemberMapper">
+<mapper namespace="com.elixter.kopring.mapper.member.MemberMapper">
 
     <insert id="save" parameterType="Member" useGeneratedKeys="true" keyProperty="id" keyColumn="id">
         INSERT INTO members(name, login_id, password, email, role)
