+++ kopring/src/test/kotlin/com/elixter/kopring/mapper/member/MemberMapperTest.kt
@@ -1,6 +1,6 @@
-package com.elixter.kopring.mapper
+package com.elixter.kopring.mapper.member
 
-import com.elixter.kopring.domain.Member
+import com.elixter.kopring.domain.member.Member
 import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
 import org.slf4j.LoggerFactory
